requirements.txt
pandas
pygetwindow
pynput
pyyaml
pyarrow
